Mark Phabricator as having remote hooks installed
Summary:
The name of this config option is super confusing since it really just means "do
not implicitly 'mark-committed' when running 'arc commit' or 'arc amend'" but
that's kind of a lot of text. Not sure what best to do about this.

If this looks ok, I'll also make this change to libphutil, diviner, arcanist and
javelin.

Test Plan:
will amend

Reviewed By: tuomaspelkonen
Reviewers: tuomaspelkonen, aran, jungejason
CC: aran, epriestley, tuomaspelkonen
Differential Revision: 245