Fix pastebin mimetypes.
Summary:
PhabricatorFile() was setting the mimetype based on extension, meaning that you couldn't view the plain-text file if you saved a file, for example, as a .php. It would set the mimetype to "text/x-php; charset=us-ascii". In this commit, I force the mimetype to text/plain.

Test Plan:
Tried pasting a new file and was able to both see it via the pastebin viewer and in plain-text via File.

Reviewers:
epriestley

CC:

Differential Revision: 429